Move base u64 type to usize

+++ examples/fastfile_fastread.rs
@@ -9,15 +9,15 @@ fn main() {
         .open()
         .expect("Failed to open path as FastFile");
 
-    let mut bytes_read = 0u64;
-    let mut sum = 0u64;
+    let mut bytes_read = 0usize;
+    let mut sum = 0usize;
     loop {
         let buf = ffr.read().expect("Failed to fastread file");
         if buf.is_empty() {
             break;
         };
-        bytes_read += buf.len() as u64;
-        sum += buf.iter().map(|x| u64::from(*x)).sum::<u64>();
+        bytes_read += buf.len();
+        sum += buf.iter().map(|x| usize::from(*x)).sum::<usize>();
     }
     assert_eq!(bytes_read, ffr.size(), "Read bytes differ from file size");
     println!("Bytes read: {}, expected: {}, sum: {}", bytes_read, ffr.size(), sum);
